Rename 01 04 2024 microsecond timeline to 01 04 2024 microsecond timeline (see comment!)
You will have to store the data to data file before the next trial, or the data isn't going to be saved due to never changing the excel name from the design of the code. 
Plus don't open a same name excel while running this code, else it won't run!